refactor: modularize character tabs
