+++ components/character-tabs/core-stats/CoreStatsTab.tsx
