+++ components/character-tabs/equipment/EquipmentTab.tsx
