+++ components/character-tabs/powers/PowersTab.tsx
