+++ components/character-tabs/rulings/RulingsTab.tsx
