+++ components/character-tabs/social/SocialTab.tsx
